feat: se suben cambios de busqeuda activa en la esncuesta

+++ database/migrations/2025_09_05_174710_add_order_is_requeride_to_survey_questions.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('survey_questions', function (Blueprint $table) {
-             $table->integer('order')->nullable()->after('id');
+             $table->decimal('order')->nullable()->after('id');
              $table->boolean('is_required')->default(false)->after('order');
         });
     }
